@@ -7,27 +7,26 @@
     "@fortawesome/free-brands-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.16",
-    "@lingui/detect-locale": "^3.13.0",
-    "@lingui/react": "^3.13.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.24.0",
-    "bulma": "^0.9.3",
+    "@lingui/detect-locale": "^3.15.0",
+    "@lingui/react": "^3.15.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.2.2",
+    "bulma": "^0.9.4",
     "bulma-toast": "^2.4.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bulma-components": "^4.0.7",
-    "react-dom": "^17.0.2",
-    "react-helmet-async": "^1.2.2",
-    "react-hook-form": "^7.22.2",
-    "react-intersection-observer": "^8.33.1",
-    "react-medium-image-zoom": "^4.3.7",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
+    "react-hook-form": "^7.41.2",
+    "react-intersection-observer": "^9.4.1",
+    "react-medium-image-zoom": "^5.1.2",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "lingui compile && react-scripts start",
     "build": "lingui compile && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
@@ -44,7 +43,7 @@
   },
   "browserslist": {
     "production": [
-      ">1%",
+      ">2%",
       "not dead",
       "not op_mini all"
     ],
@@ -55,9 +54,9 @@
     ]
   },
   "devDependencies": {
-    "@lingui/cli": "^3.13.0",
-    "@lingui/macro": "^3.13.0",
-    "sass": "^1.45.0",
-    "standard": "^16.0.4"
+    "@lingui/cli": "^3.15.0",
+    "@lingui/macro": "^3.15.0",
+    "sass": "^1.51.0",
+    "standard": "^17.0.0"
   }
 }
