@@ -1,12 +1,14 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.scss'
 import App from './components/App'
 import { HelmetProvider } from 'react-helmet-async'
 import { AuthProvider } from './hooks/auth'
 import TranslationProvider from './i18n'
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'))
+
+root.render(
   <React.StrictMode>
     <HelmetProvider>
       <AuthProvider>
@@ -15,6 +17,5 @@ ReactDOM.render(
         </TranslationProvider>
       </AuthProvider>
     </HelmetProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
